Implemented an interface utilised by Obfuscate and Report
