+++ src/main/java/com/mindlinksoft/recruitment/mychat/Constructs/Conversation.java
@@ -1,4 +1,4 @@
-package com.mindlinksoft.recruitment.mychat.Objects;
+package com.mindlinksoft.recruitment.mychat.Constructs;
 
 import java.util.Collection;
