+++ src/main/java/com/mindlinksoft/recruitment/mychat/Constructs/ConversationExporterConfiguration.java
@@ -1,4 +1,4 @@
-package com.mindlinksoft.recruitment.mychat.Objects;
+package com.mindlinksoft.recruitment.mychat.Constructs;
 
 public final class ConversationExporterConfiguration {
