+++ src/main/java/com/mindlinksoft/recruitment/mychat/Constructs/FlagContainer.java
@@ -1,4 +1,4 @@
-package com.mindlinksoft.recruitment.mychat.Objects;
+package com.mindlinksoft.recruitment.mychat.Constructs;
 
 public class FlagContainer {
